cleanup(many): Remove feature flags for reset password with code
Because:

* Reset password with code is now fully rolled out

This commit:

* Remove functional tests for reset password with link
* Remove config check for resetPasswordWithCode flag in functional-tests
* Remove content-server resetPasswordWithCode feature flag/env
* Remove auth-server passwordForgotOtp.enabled config

Closes #FXA-9728, FXA-10387
