+++ packages/functional-tests/tests/resetPassword/resetPasswordRecoveryKey.spec.ts
@@ -2,50 +2,34 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { expect, test } from '../../../lib/fixtures/standard';
-
-const HINT = 'secret key location';
+import { expect, test } from '../../lib/fixtures/standard';
+import { SettingsPage } from '../../pages/settings';
+import { RecoveryKeyPage } from '../../pages/settings/recoveryKey';
+import { SigninPage } from '../../pages/signin';
 
 test.describe('severity-1 #smoke', () => {
   test.describe('recovery key react', () => {
-    test.beforeEach(async ({ pages: { configPage } }) => {
-      // Ensure that the react reset password route feature flag is enabled
-      const config = await configPage.getConfig();
-      test.skip(
-        config.featureFlags.resetPasswordWithCode === true,
-        'see FXA-9728, remove these tests'
-      );
-    });
-
     test('can reset password with recovery key', async ({
       target,
       page,
-      pages: { settings, recoveryKey, signin, resetPassword },
+      pages: { recoveryKey, resetPassword, settings, signin },
       testAccountTracker,
     }) => {
       const credentials = await testAccountTracker.signUp();
       const newPassword = testAccountTracker.generatePassword();
 
-      await signin.goto();
-      await signin.fillOutEmailFirstForm(credentials.email);
-      await signin.fillOutPasswordForm(credentials.password);
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Not Set');
-
-      await settings.recoveryKey.createButton.click();
-
-      const key = await recoveryKey.createRecoveryKey(
+      await signinAccount(
+        credentials.email,
         credentials.password,
-        HINT
+        settings,
+        signin
       );
 
-      // Verify status as 'enabled'
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Enabled');
-
-      // Ensure password reset occurs with no session token available
-      await signin.clearCache();
+      const key = await enableRecoveryKey(
+        credentials.password,
+        recoveryKey,
+        settings
+      );
 
       // Stash original encryption keys to be verified later
       const accountData = await target.authClient.sessionReauth(
@@ -57,7 +41,6 @@ test.describe('severity-1 #smoke', () => {
           reason: 'recovery_key',
         }
       );
-
       expect(accountData.keyFetchToken).toBeDefined();
       expect(accountData.unwrapBKey).toBeDefined();
 
@@ -69,18 +52,25 @@ test.describe('severity-1 #smoke', () => {
       await resetPassword.goto();
 
       await resetPassword.fillOutEmailForm(credentials.email);
-      await expect(resetPassword.confirmResetPasswordHeading).toBeVisible();
 
-      const link = await target.emailClient.getRecoveryLink(credentials.email);
-      await page.goto(link);
+      const code = await target.emailClient.getResetPasswordCode(
+        credentials.email
+      );
+
+      await resetPassword.fillOutResetPasswordCodeForm(code);
+
       await resetPassword.fillOutRecoveryKeyForm(key);
       await resetPassword.fillOutNewPasswordForm(newPassword);
-      credentials.password = newPassword;
+
       // After using a recovery key to reset password, expect to be prompted to create a new one
+      await expect(resetPassword.generateRecoveryKeyButton).toBeVisible();
       await resetPassword.generateRecoveryKeyButton.click();
+
+      // TODO in FXA-7904 - Verify that a new recovery key is generated without needing to sign in again
+      // not currently implemented
       await page.waitForURL(/settings\/account_recovery/);
 
-      // Attempt to login with new password
+      // Attempt to sign in with new password
       const { sessionToken } = await target.authClient.signIn(
         credentials.email,
         newPassword
@@ -103,45 +93,81 @@ test.describe('severity-1 #smoke', () => {
         newAccountData.unwrapBKey as string
       );
       expect(originalEncryptionKeys).toEqual(newEncryptionKeys);
+
+      // update password for cleanup function
+      credentials.password = newPassword;
     });
 
     test('forgot password has account recovery key but skip using it', async ({
       target,
-      page,
       pages: { settings, recoveryKey, resetPassword, signin },
       testAccountTracker,
     }) => {
       const credentials = await testAccountTracker.signUp();
+      const newPassword = testAccountTracker.generatePassword();
 
-      await signin.goto();
-      await signin.fillOutEmailFirstForm(credentials.email);
-      await signin.fillOutPasswordForm(credentials.password);
+      await signinAccount(
+        credentials.email,
+        credentials.password,
+        settings,
+        signin
+      );
 
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Not Set');
-
-      await settings.recoveryKey.createButton.click();
-      await recoveryKey.createRecoveryKey(credentials.password, 'hint');
-
-      await expect(settings.settingsHeading).toBeVisible();
-      await expect(settings.recoveryKey.status).toHaveText('Enabled');
+      await enableRecoveryKey(credentials.password, recoveryKey, settings);
 
       await resetPassword.goto();
 
       await resetPassword.fillOutEmailForm(credentials.email);
-      const link = await target.emailClient.getRecoveryLink(credentials.email);
-      await page.goto(link);
+      const code = await target.emailClient.getResetPasswordCode(
+        credentials.email
+      );
+
+      await resetPassword.fillOutResetPasswordCodeForm(code);
       await resetPassword.forgotKeyLink.click();
-      await resetPassword.fillOutNewPasswordForm(credentials.password);
+      await expect(resetPassword.dataLossWarning).toBeVisible();
+      await resetPassword.fillOutNewPasswordForm(newPassword);
 
       await expect(
         resetPassword.passwordResetConfirmationHeading
       ).toBeVisible();
 
-      await settings.goto();
+      await signinAccount(credentials.email, newPassword, settings, signin);
 
-      await expect(settings.settingsHeading).toBeVisible();
       await expect(settings.recoveryKey.status).toHaveText('Not Set');
+
+      // update password for cleanup function
+      credentials.password = newPassword;
     });
   });
+
+  async function signinAccount(
+    email: string,
+    password: string,
+    settings: SettingsPage,
+    signin: SigninPage
+  ): Promise<void> {
+    await signin.goto();
+    await signin.fillOutEmailFirstForm(email);
+    await signin.fillOutPasswordForm(password);
+
+    await expect(settings.settingsHeading).toBeVisible();
+  }
+
+  async function enableRecoveryKey(
+    password: string,
+    recoveryKey: RecoveryKeyPage,
+    settings: SettingsPage
+  ): Promise<string> {
+    await expect(settings.recoveryKey.status).toHaveText('Not Set');
+
+    await settings.recoveryKey.createButton.click();
+    const key = await recoveryKey.createRecoveryKey(password, 'hint');
+
+    await expect(settings.settingsHeading).toBeVisible();
+    await expect(settings.recoveryKey.status).toHaveText('Enabled');
+
+    await settings.signOut();
+
+    return key;
+  }
 });
