+++ packages/fxa-content-server/server/lib/beta-settings.js
@@ -102,7 +102,6 @@ const settingsConfig = {
   },
   featureFlags: {
     keyStretchV2: config.get('featureFlags.keyStretchV2'),
-    resetPasswordWithCode: config.get('featureFlags.resetPasswordWithCode'),
     recoveryCodeSetupOnSyncSignIn: config.get(
       'featureFlags.recoveryCodeSetupOnSyncSignIn'
     ),
