+++ packages/fxa-content-server/server/lib/routes/get-index.js
@@ -52,9 +52,6 @@ module.exports = function (config) {
   const FEATURE_FLAGS_FXA_STATUS_ON_SETTINGS = config.get(
     'featureFlags.sendFxAStatusOnSettings'
   );
-  const FEATURE_FLAGS_RESET_PWD_WITH_CODE = config.get(
-    'featureFlags.resetPasswordWithCode'
-  );
   const FEATURE_FLAGS_RECOVERY_CODE_SETUP_ON_SYNC_SIGN_IN = config.get(
     'featureFlags.recoveryCodeSetupOnSyncSignIn'
   );
@@ -116,7 +113,6 @@ module.exports = function (config) {
     brandMessagingMode: BRAND_MESSAGING_MODE,
     featureFlags: {
       sendFxAStatusOnSettings: FEATURE_FLAGS_FXA_STATUS_ON_SETTINGS,
-      resetPasswordWithCode: FEATURE_FLAGS_RESET_PWD_WITH_CODE,
       recoveryCodeSetupOnSyncSignIn:
         FEATURE_FLAGS_RECOVERY_CODE_SETUP_ON_SYNC_SIGN_IN,
     },
